new way to obtain alignability accuracy and precision now it is obtained the alignability of the projection direction. So the mirror is not obtained

+++ software/em/xmipp/libraries/reconstruction/angular_accuracy_pca.h
@@ -52,9 +52,9 @@ public:
 
     PCAMahalanobisAnalyzer pca;
 
-	size_t newXdim;
+	int newXdim;
 
-	size_t newYdim;
+	int newYdim;
